pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "semp-workflow"
version = "0.2.0"
description = "Standalone CI/CD Workflow Automation for Solace SEMP"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "requests>=2.31",
    "click>=8.1",
    "jinja2>=3.1",
    "colorama>=0.4",
    "urllib3>=2.0",
]

[project.scripts]
semp-workflow = "semp_workflow.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=7.4",
    "pytest-mock>=3.11",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "integration: requires a live Solace broker (set SEMP_HOST, SEMP_USERNAME, SEMP_PASSWORD, SEMP_MSG_VPN)",
]

[tool.setuptools.packages.find]
where = ["src"]
